Fix network coloring by preventing uglify function name mangling
After processing the ES6 code with Babel, a class' constructor is just a
regular function whose name is later mangled by uglify such that calls
to obj.constructor.name fail because they return the mangled function
name instead of the class name. At the moment, certain parts of the
neural network are colored based on the class name of the nodes in the
graph.