Add training and training UI through refactoring of NN property store
Properties of nodes and edges are now stored separately by the classes
that need them. The network only stores the topology. Properties are
mapped to nodes and edges in the network via node and edge ids.

This allows for much better separation of concerns of the network model,
the training, user provided input and temporary properties needed
for the network visualization. This is mostly due to the fact that
properties are now stored in a simple map that allows for easy cloning
and addition of properties. The previous store in form of a directed
acyclic graph was much harder to work with. Changes to the
properties need to be tracked by different parts of the code that
otherwise don't have much to do with each other. Properties are now
tied much more to the places they belong to and can be hidden from other
parts of the code when necessary.