Connect slider and level loader
Since a couple of views add event listeners to the DOM, I added a
DOMEventManager that keeps track of listeners and detached them from the
DOM when needed to allow for proper garbage collection.

If the slider is navigated in rapid succession, the level loader will
always wait until the previous load request is complete (independent of
its success) before loading the next level.